Fixes wrong updates of file modification time (mtime) when only access time (atime) should be updated
Using file:write_file_info/2 to update atime, results in updating both atime and mtime. mtime is the last time the file was written, so it shouldn't be updated when atime (last time the file was read).
The fix changes that behavior such that when mtime field is not give, it's not updated.